requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bestconfig==1.3.5
certifi==2022.9.14
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
Pillow==9.2.0
pydantic==1.10.2
Pygments==2.13.0
pytz==2022.2.1
PyYAML==6.0
typing_extensions==4.3.0
yarl==1.8.1
